backend/requirements.txt
apscheduler
asgiref==3.7.2
boto3==1.35.71
botocore==1.35.71
cachetools==5.3.1
certifi==2024.8.30
charset-normalizer==3.3.0
Django==4.2.5
django-apscheduler
django-cors-headers==4.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
google-api-core==2.23.0
google-auth==2.26.1
google-cloud==0.34.0
google-cloud-documentai==2.33.0
google-cloud-storage==2.18.2
google-cloud-vision==3.4.4
googleapis-common-protos==1.60.0
grpcio==1.67.0
grpcio-status==1.57.0
gunicorn==23.0.0
idna==3.4
inflection==0.5.1
joblib==1.4.2
mysqlclient==2.2.4
numpy==2.1.1
packaging==24.1
Pillow==10.4.0
proto-plus==1.22.3
protobuf==4.25.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.2.0
six==1.16.0
stripe==11.1.1
sqlparse==0.4.4
threadpoolctl==3.5.0
typing_extensions==4.8.0
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.0.5
whitenoise==6.7.0